perf: don't load ResourceView on home page
